requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
openai==1.3.6
spacy==3.6.1
python-dotenv==1.0.0
gunicorn==21.2.0